Swapping internal dependencies to aiohttp
